refactor: Extract input submodule in ui module
