+++ src/loveletter_cli/ui/__init__.py
@@ -1,2 +1,3 @@
 from .board import *
+from .input import *
 from .misc import *
